+++ pkg/pull/model.go
@@ -21,96 +21,82 @@ import (
 	"encoding/json"
 
 	over "github.com/Trendyol/overlog"
+	"github.com/cgi-fr/jsonline/pkg/jsonline"
 	"github.com/rs/zerolog/log"
 )
 
-// Table from which to pull data.
-type Table interface {
-	Name() string
-	PrimaryKey() []string
-	Columns() ColumnList
-	export(Row) ExportableRow
+type Cardinality bool
+
+const (
+	Many Cardinality = true
+	One  Cardinality = false
+)
+
+type (
+	TableName    string
+	RelationName string
+)
+
+type Column struct {
+	Name   string
+	Export string
 }
 
-// ColumnList is a list of columns.
-type ColumnList interface {
-	Len() uint
-	Contains(string) bool
-	Column(idx uint) Column
+type Table struct {
+	Name    TableName
+	Keys    []string
+	Columns []Column
 
-	add(Column) ColumnList
+	template jsonline.Template
 }
 
-// Column of a table.
-type Column interface {
-	Name() string
-	Export() string
+type RelationTip struct {
+	Table Table
+	Keys  []string
 }
 
-// Relation between two tables.
-type Relation interface {
-	Name() string
-	Parent() Table
-	Child() Table
-	ParentKey() []string
-	ChildKey() []string
-	OppositeOf(tablename string) Table
+type Relation struct {
+	Name        RelationName
+	Cardinality Cardinality
+	Local       RelationTip
+	Foreign     RelationTip
 }
 
-// RelationList is a list of relations.
-type RelationList interface {
-	Len() uint
-	Relation(idx uint) Relation
+type RelationSet []Relation
+
+type Plan struct {
+	Relations  RelationSet
+	Components map[TableName]uint // <= could be deduced from relations with tarjan algorithm
 }
 
-// Cycle is a list of relations.
-type Cycle interface {
-	RelationList
+type Graph struct {
+	Relations  map[TableName]RelationSet
+	Components map[TableName]uint
+	Cached     map[TableName]bool
 }
 
-// CycleList is a list of cycles.
-type CycleList interface {
-	Len() uint
-	Cycle(idx uint) Cycle
+type Row map[string]interface{}
+
+type RowSet []Row
+
+type DataSet map[TableName]RowSet
+
+type Filter struct {
+	Limit    uint
+	Values   Row
+	Where    string
+	Distinct bool
 }
 
-// Step group of follows to perform.
-type Step interface {
-	Index() uint
-	Entry() Table
-	Follow() Relation
-	Relations() RelationList
-	Cycles() CycleList
-	NextSteps() StepList
+// ExportedRow is a row but with keys ordered and values in export format for jsonline.
+type ExportedRow struct {
+	jsonline.Row
 }
 
-// StepList list of steps to perform.
-type StepList interface {
-	Len() uint
-	Step(uint) Step
-}
+func (er ExportedRow) GetOrNil(key string) interface{} {
+	v, _ := er.Get(key)
 
-// Plan of the puller process.
-type Plan interface {
-	InitFilter() Filter
-	Steps() StepList
-}
-
-// Filter applied to data tables.
-type Filter interface {
-	Limit() uint
-	Values() Row
-	Where() string
-	Distinct() bool
-}
-
-// Error is the error type returned by the domain
-type Error struct {
-	Description string
-}
-
-func (e *Error) Error() string {
-	return e.Description
+	return v
 }
 
 type ExecutionStats interface {
@@ -125,22 +111,6 @@ type stats struct {
 	FiltersCount      int            `json:"filtersCount"`
 }
 
-// Reset all statistics to zero
-func Reset() {
-	over.MDC().Set("stats", &stats{LinesPerStepCount: map[string]int{}})
-}
-
-// Compute current statistics and give a snapshot
-func Compute() ExecutionStats {
-	value, exists := over.MDC().Get("stats")
-	if stats, ok := value.(ExecutionStats); exists && ok {
-		return stats
-	}
-	log.Warn().Msg("Unable to compute statistics")
-	return &stats{}
-}
-
-// Exports statistics to readable json format
 func (s *stats) ToJSON() []byte {
 	b, err := json.Marshal(s)
 	if err != nil {
@@ -167,7 +137,6 @@ func IncFiltersCount() {
 	stats.FiltersCount++
 }
 
-// Compute current statistics and give a snapshot
 func getStats() *stats {
 	value, exists := over.MDC().Get("stats")
 	if stats, ok := value.(*stats); exists && ok {
@@ -175,4 +144,17 @@ func getStats() *stats {
 	}
 	log.Warn().Msg("Statistics uncorrectly initialized")
 	return &stats{}
+}
+
+func Compute() ExecutionStats {
+	value, exists := over.MDC().Get("stats")
+	if stats, ok := value.(ExecutionStats); exists && ok {
+		return stats
+	}
+	log.Warn().Msg("Unable to compute statistics")
+	return &stats{}
+}
+
+func Reset() {
+	over.MDC().Set("stats", &stats{LinesPerStepCount: map[string]int{}})
 }