+++ CHANGELOG.md
@@ -18,6 +18,7 @@ Types of changes
 
 - `Added` configuration of exported columns in tables.yaml, see issue #33 for more information.
 - `Added` MariaDB/MySQL support (thanks to @joaking85).
+- `Added` auto-select columns required by a relation but not exported in tables.yaml.
 - `Changed` order of keys in output JSON lines will be alphabetical when pulling (without configuration in tables.yaml).
 - `Fixed` limit keyword on DB2 dialect.